Refactored to allow rules to scale
