+++ .gitignore
@@ -4,3 +4,4 @@ node_modules/*
 package-lock.json
 .vscode/*
 secrets.ts
+node_modules
