+++ functions/authHelpers.ts
