+++ functions/fa-etiquette-create-subs/index.ts
@@ -9,7 +9,7 @@ const httpTrigger: AzureFunction = async function (context: Context, req: HttpRe
 
     // TODO: Get user ID from querystring
     
-    await subscriptionHelper.addSubscription('48e8a1ab-0d3a-4f9b-b200-e9e9d1437a2b', context);
+    await subscriptionHelper.addEventSubscription('48e8a1ab-0d3a-4f9b-b200-e9e9d1437a2b', context);
     context.log("About to initalise user");
     await userHelper.initialiseUserExtension('48e8a1ab-0d3a-4f9b-b200-e9e9d1437a2b', context);
     context.log("Subscription created");
