+++ functions/graph-helpers.ts
