+++ functions/readme.md
@@ -20,3 +20,10 @@ The rules are based on what can be triggered via Graph events.
 - Groups
 - Teams Calls
 - Teams Chat
+
+## How does it run?
+
+Configure app registration and store details in secrets.ts
+Run "ngrok http 7071 host-header="localhost:7071"
+Run "npm run dev"
+
