+++ functions/secrets.ts.sample
